requirements.txt
asgiref==3.9.1
Django==4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
drf-spectacular
pytest
pytest-django
gunicorn
psycopg2-binary
dj-database-url
whitenoise
python-dotenv